Remove dependency on aaa
We are currently undergoing work in AAA project to consolidate on the
features-aaa feature repository and not provide features-aaa-shiro. Part
of this work includes evaluation of downstream dependencies use of
features-aaa-shiro;  from this standpoint usc looks like an atypical
use case.  I.e., most projects that include features-aaa-shiro (or
features-aaa) do so in order to provide programatic reference to
AAAShiroFilter from within web.xml.  usc does not provide such JAX-RS
interface from its project.

aaa is enabled as a part of RESTCONF implementations through the
NETCONF feature odl-restconf;  is this sufficient for usc?  I will point
out that the only dependency for AAA is currently in the karaf bundle,
which implies that the existing features really don't depend on it,
as the karaf bundle is neither installed nor deployed by default.

Change-Id: Iafad9a5a065ebe4d311e5981c507469c77a6ef24
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>