Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue-ID: RELENG-86
Change-Id: I030f9954fdc11926a4b9a2ee9ac1f0061279486f
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>